Handle trying to delete a Greeting that does not exist.
The controllers need to check that a Greeting exists and throw an
exception if not.

+++ src/main/java/hello/controllers/GreetingController.java
@@ -87,6 +87,10 @@ public class GreetingController {
 
 	@DeleteMapping("/{id}")
 	public String deleteGreeting(@PathVariable("id") long id, RedirectAttributes redirectAttrs) {
+		if (!greetingService.existsById(id)) {
+			throw new GreetingNotFoundException(id);
+		}
+
 		greetingService.deleteById(id);
 		redirectAttrs.addFlashAttribute("ok_message", "Greeting deleted.");
